# docs/swagger.yaml
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: LL
    url: http://www.swagger.io/support
  description: 基于Gin、Gorm框架的在线练习系统
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 在线练习系统
  version: "1.0"
paths:
  /login:
    post:
      parameters:
        - description: 用户名称username
          in: formData
          name: username
          type: string
        - description: 用户密码password
          in: formData
          name: password
          type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 用户登录
      tags:
        - 公共接口
  /problem-list:
    get:
      parameters:
        - description: page
          in: query
          name: page
          type: integer
        - description: size
          in: query
          name: size
          type: integer
        - description: keyword
          in: query
          name: keyword
          type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 问题列表
      tags:
        - 公共接口
  /rank-list:
    get:
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 用户排行榜
      tags:
        - 公共接口
  /register:
    post:
      parameters:
        - description: 邮箱
          in: formData
          name: email
          required: true
          type: string
        - description: 验证码
          in: formData
          name: code
          required: true
          type: string
        - description: 名字
          in: formData
          name: name
          required: true
          type: string
        - description: 密码
          in: formData
          name: password
          required: true
          type: string
        - description: 手机号
          in: formData
          name: phone
          type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 用户注册
      tags:
        - 公共接口
  /send-code:
    post:
      parameters:
        - description: email
          in: formData
          name: email
          required: true
          type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 发送验证码
      tags:
        - 公共接口
  /submit-list:
    get:
      parameters:
        - description: page
          in: query
          name: page
          type: integer
        - description: size
          in: query
          name: size
          type: integer
        - description: problem_identity
          in: query
          name: problem_identity
          type: string
        - description: user_identity
          in: query
          name: user_identity
          type: string
        - description: status
          in: query
          name: status
          type: integer
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 提交列表
      tags:
        - 公共接口
  /user-list:
    get:
      parameters:
        - description: 用户标识identity
          in: query
          name: identity
          type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 用户详情
      tags:
        - 公共接口
swagger: "2.0"